% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adae}
\alias{adam_adae}
\title{adam_adae}
\format{
a data frame with 1191 rows and 55 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adae
}
\description{
ADaM adae data frame
}
\keyword{datasets}
