% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adlbc}
\alias{adam_adlbc}
\title{adam_adlbc}
\format{
a data frame with 74264 rows and 46 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adlbc
}
\description{
ADaM adlbc data frame
}
\keyword{datasets}
