% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adlbh}
\alias{adam_adlbh}
\title{adam_adlbh}
\format{
a data frame with 49932 rows and 46 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adlbh
}
\description{
ADaM adlbh data frame
}
\keyword{datasets}
