% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adlbhy}
\alias{adam_adlbhy}
\title{adam_adlbhy}
\format{
a data frame with 9954 rows and 43 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adlbhy
}
\description{
ADaM adlbhy data frame
}
\keyword{datasets}
