% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adqsadas}
\alias{adam_adqsadas}
\title{adam_adqsadas}
\format{
a data frame with 12463 rows and 40 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adqsadas
}
\description{
ADaM adqsadas data frame
}
\keyword{datasets}
