% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adqscibc}
\alias{adam_adqscibc}
\title{adam_adqscibc}
\format{
a data frame with 730 rows and 36 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adqscibc
}
\description{
ADaM adqscibc data frame
}
\keyword{datasets}
