% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adsl}
\alias{adam_adsl}
\title{adam_adsl}
\format{
a data frame with 254 rows and 48 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adsl
}
\description{
ADaM adsl data frame
}
\keyword{datasets}
