% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adtte}
\alias{adam_adtte}
\title{adam_adtte}
\format{
a data frame with 254 rows and 26 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adtte
}
\description{
ADaM adtte data frame
}
\keyword{datasets}
