% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_advs}
\alias{adam_advs}
\title{adam_advs}
\format{
a data frame with 32139 rows and 34 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_advs
}
\description{
ADaM advs data frame
}
\keyword{datasets}
