% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ae}
\alias{ae}
\title{ae}
\format{
data.frame
'data.frame':	1191 obs. of  35 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "AE" "AE" "AE" "AE" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1023" ...
$ AESEQ   : int  1 2 3 3 1 2 4 1 2 1 ...
$ AESPID  : chr  "E07" "E08" "E06" "E10" ...
$ AETERM  : chr  "APPLICATION SITE ERYTHEMA" "APPLICATION SITE PRURITUS" "DIARRHOEA" "ATRIOVENTRICULAR BLOCK SECOND DEGREE" ...
$ AELLT   : chr  "APPLICATION SITE REDNESS" "APPLICATION SITE ITCHING" "DIARRHEA" "AV BLOCK SECOND DEGREE" ...
$ AELLTCD : logi  NA NA NA NA NA NA ...
$ AEDECOD : chr  "APPLICATION SITE ERYTHEMA" "APPLICATION SITE PRURITUS" "DIARRHOEA" "ATRIOVENTRICULAR BLOCK SECOND DEGREE" ...
$ AEPTCD  : logi  NA NA NA NA NA NA ...
$ AEHLT   : chr  "HLT_0617" "HLT_0317" "HLT_0148" "HLT_0415" ...
$ AEHLTCD : logi  NA NA NA NA NA NA ...
$ AEHLGT  : chr  "HLGT_0152" "HLGT_0338" "HLGT_0588" "HLGT_0086" ...
$ AEHLGTCD: logi  NA NA NA NA NA NA ...
$ AEBODSYS: chr  "GENERAL DISORDERS AND ADMINISTRATION SITE CONDITIONS" "GENERAL DISORDERS AND ADMINISTRATION SITE CONDITIONS" "GASTROINTESTINAL DISORDERS" "CARDIAC DISORDERS" ...
$ AEBDSYCD: logi  NA NA NA NA NA NA ...
$ AESOC   : chr  "GENERAL DISORDERS AND ADMINISTRATION SITE CONDITIONS" "GENERAL DISORDERS AND ADMINISTRATION SITE CONDITIONS" "GASTROINTESTINAL DISORDERS" "CARDIAC DISORDERS" ...
$ AESOCCD : logi  NA NA NA NA NA NA ...
$ AESEV   : chr  "MILD" "MILD" "MILD" "MILD" ...
$ AESER   : chr  "N" "N" "N" "N" ...
$ AEACN   : logi  NA NA NA NA NA NA ...
$ AEREL   : chr  "PROBABLE" "PROBABLE" "REMOTE" "POSSIBLE" ...
$ AEOUT   : chr  "NOT RECOVERED/NOT RESOLVED" "NOT RECOVERED/NOT RESOLVED" "RECOVERED/RESOLVED" "NOT RECOVERED/NOT RESOLVED" ...
$ AESCAN  : chr  "N" "N" "N" "N" ...
$ AESCONG : chr  "N" "N" "N" "N" ...
$ AESDISAB: chr  "N" "N" "N" "N" ...
$ AESDTH  : chr  "N" "N" "N" "N" ...
$ AESHOSP : chr  "N" "N" "N" "N" ...
$ AESLIFE : chr  "N" "N" "N" "N" ...
$ AESOD   : chr  "N" "N" "N" "N" ...
$ AEDTC   : chr  "2014-01-16" "2014-01-16" "2014-01-16" "2012-08-27" ...
$ AESTDTC : chr  "2014-01-03" "2014-01-03" "2014-01-09" "2012-08-26" ...
$ AEENDTC : chr  NA NA "2014-01-11" NA ...
$ AESTDY  : int  2 2 8 22 3 3 3 3 21 58 ...
$ AEENDY  : int  NA NA 10 NA 26 NA 26 NA NA NA ...
}
\usage{
ae
}
\description{
ae
}
\keyword{datasets}
