% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AirportCodes_20170404}
\alias{AirportCodes_20170404}
\title{IATA and ICAO Airport Codes from 2017-04-04}
\format{A data frame with 30908 rows and 2 variables:
\describe{
  \item{IATA_Code}{3-letter IATA codes}
  \item{ICAO_Code}{4-letter ICAO codes}
}}
\usage{
AirportCodes_20170404
}
\description{
A list of global ICAO and IATA airport code from 2017-04-04.
Useful for converting between the two since airport-related data is often available in one or the other format.
}
\keyword{datasets}
