% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AirportTimezones_20170404}
\alias{AirportTimezones_20170404}
\title{Airport Timezones from 2017-04-04}
\format{A data frame with 3527 rows and 2 variables:
\describe{
  \item{Code}{4-letter ICAO codes}
  \item{Zulu_Offset}{Integer representing hours offset from Zulu (UTC)}
}}
\usage{
AirportTimezones_20170404
}
\description{
A list of airport timezones from 2017-04-04.
}
\keyword{datasets}
