% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_trial_dataframe.R
\docType{data}
\name{clinical_trial_dataframe}
\alias{clinical_trial_dataframe}
\title{Clinical data frames}
\format{
A data frame with 16893 rows and 25 variables:
\describe{
\item{\code{USUBJID}}{character COLUMN_DESCRIPTION}
\item{\code{SITE}}{character COLUMN_DESCRIPTION}
\item{\code{SITEID}}{integer COLUMN_DESCRIPTION}
\item{\code{AGE}}{integer COLUMN_DESCRIPTION}
\item{\code{SEX}}{character COLUMN_DESCRIPTION}
\item{\code{RACE}}{character COLUMN_DESCRIPTION}
\item{\code{ARM}}{character COLUMN_DESCRIPTION}
\item{\code{ARMCD}}{character COLUMN_DESCRIPTION}
\item{\code{SBJTSTAT}}{character COLUMN_DESCRIPTION}
\item{\code{RFSTDTC}}{character COLUMN_DESCRIPTION}
\item{\code{RFENDTC}}{character COLUMN_DESCRIPTION}
\item{\code{RFENDY}}{integer COLUMN_DESCRIPTION}
\item{\code{SAFFL}}{character COLUMN_DESCRIPTION}
\item{\code{SAFFN}}{integer COLUMN_DESCRIPTION}
\item{\code{VISIT}}{character COLUMN_DESCRIPTION}
\item{\code{VISITNUM}}{double COLUMN_DESCRIPTION}
\item{\code{DT}}{character COLUMN_DESCRIPTION}
\item{\code{DY}}{integer COLUMN_DESCRIPTION}
\item{\code{CAT}}{character COLUMN_DESCRIPTION}
\item{\code{TEST}}{character COLUMN_DESCRIPTION}
\item{\code{STRESU}}{character COLUMN_DESCRIPTION}
\item{\code{STRESN}}{double COLUMN_DESCRIPTION}
\item{\code{STNRLO}}{double COLUMN_DESCRIPTION}
\item{\code{STNRHI}}{double COLUMN_DESCRIPTION}
\item{\code{VISITN}}{double COLUMN_DESCRIPTION}
}
}
\usage{
clinical_trial_dataframe
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
