% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{cm}
\alias{cm}
\title{cm}
\format{
data.frame
'data.frame':	7510 obs. of  21 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "CM" "CM" "CM" "CM" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ CMSEQ   : int  1 5 9 13 18 23 28 33 38 43 ...
$ CMSPID  : int  1 1 1 1 1 1 1 1 1 1 ...
$ CMTRT   : chr  "ASPIRIN" "ASPIRIN" "ASPIRIN" "ASPIRIN" ...
$ CMDECOD : chr  "ACETYLSALICYLIC ACID" "ACETYLSALICYLIC ACID" "ACETYLSALICYLIC ACID" "ACETYLSALICYLIC ACID" ...
$ CMINDC  : chr  NA NA NA NA ...
$ CMCLAS  : chr  "NERVOUS SYSTEM" "NERVOUS SYSTEM" "NERVOUS SYSTEM" "NERVOUS SYSTEM" ...
$ CMDOSE  : num  1 1 1 1 1 1 1 1 1 1 ...
$ CMDOSU  : chr  "TABLET" "TABLET" "TABLET" "TABLET" ...
$ CMDOSFRQ: chr  "PRN" "PRN" "PRN" "PRN" ...
$ CMROUTE : chr  "ORAL" "ORAL" "ORAL" "ORAL" ...
$ VISITNUM: int  1 2 3 4 5 6 7 8 9 10 ...
$ VISIT   : chr  "SCREENING 1" "SCREENING 2" "BASELINE" "WEEK 2" ...
$ VISITDY : int  -7 -1 1 14 28 30 42 56 84 112 ...
$ CMDTC   : chr  "2013-12-26" "2013-12-31" "2014-01-02" "2014-01-16" ...
$ CMSTDTC : chr  "2003" "2003" "2003" "2003" ...
$ CMENDTC : chr  NA NA NA NA ...
$ CMSTDY  : int  NA NA NA NA NA NA NA NA NA NA ...
$ CMENDY  : int  NA NA NA NA NA NA NA NA NA NA ...
}
\usage{
cm
}
\description{
cm
}
\keyword{datasets}
