% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityAltitude.R
\name{densityAltitude}
\alias{densityAltitude}
\title{Calculate Density Altitude}
\source{
National Weather Service / National Oceanic and Atmospheric Administration
Density Altitude
\url{https://www.weather.gov/media/epz/wxcalc/densityAltitude.pdf}
Temperature Conversion
\url{http://www.weather.gov/media/epz/wxcalc/tempConvert.pdf}
Pressure Conversion
\url{http://www.weather.gov/media/epz/wxcalc/pressureConversion.pdf}
}
\usage{
densityAltitude(Altimeter_inHg = 29.92, Temp_C = 15, DewPoint_C = 0)
}
\arguments{
\item{Altimeter_inHg}{Local altimeter setting measured as inches of mercury, defaults to 29.92 (ICAO standard atmosphere)}

\item{Temp_C}{Local outside air temperature in degrees Celcius, defaults to 15 (ICAO standard temperature)}

\item{DewPoint_C}{Local dew point in degrees Celcius, defaults to 0}
}
\value{
Density altitude in feet
}
\description{
Function to calculate density altitude using meteorological data available for pilots, airlines, etc.
This function uses the National Weather Service estimation
}
\examples{
densityAltitude() # Default density altitude is nearly 0 ft
densityAltitude(Altimeter_inHg = 29.50) # Density altitude increases as pressure decreases
densityAltitude(Temp_C = 30) # Density altitude increases with temperature
densityAltitude(DewPoint_C = 15) # Density altitude increases with the dew point when temperature is held constant since moist air is less dense than dry air
}
