% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{dm}
\alias{dm}
\title{dm}
\format{
data.frame
'data.frame':	306 obs. of  25 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "DM" "DM" "DM" "DM" ...
$ USUBJID : chr  "01-701-1015" "01-701-1023" "01-701-1028" "01-701-1033" ...
$ SUBJID  : int  1015 1023 1028 1033 1034 1047 1057 1097 1111 1115 ...
$ RFSTDTC : chr  "2014-01-02" "2012-08-05" "2013-07-19" "2014-03-18" ...
$ RFENDTC : chr  "2014-07-02" "2012-09-02" "2014-01-14" "2014-04-14" ...
$ RFXSTDTC: chr  "2014-01-02" "2012-08-05" "2013-07-19" "2014-03-18" ...
$ RFXENDTC: chr  "2014-07-02" "2012-09-01" "2014-01-14" "2014-03-31" ...
$ RFICDTC : logi  NA NA NA NA NA NA ...
$ RFPENDTC: chr  "2014-07-02T11:45" "2013-02-18" "2014-01-14T11:10" "2014-09-15" ...
$ DTHDTC  : chr  NA NA NA NA ...
$ DTHFL   : chr  NA NA NA NA ...
$ SITEID  : int  701 701 701 701 701 701 701 701 701 701 ...
$ AGE     : int  63 64 71 74 77 85 59 68 81 84 ...
$ AGEU    : chr  "YEARS" "YEARS" "YEARS" "YEARS" ...
$ SEX     : chr  "F" "M" "M" "M" ...
$ RACE    : chr  "WHITE" "WHITE" "WHITE" "WHITE" ...
$ ETHNIC  : chr  "HISPANIC OR LATINO" "HISPANIC OR LATINO" "NOT HISPANIC OR LATINO" "NOT HISPANIC OR LATINO" ...
$ ARMCD   : chr  "Pbo" "Pbo" "Xan_Hi" "Xan_Lo" ...
$ ARM     : chr  "Placebo" "Placebo" "Xanomeline High Dose" "Xanomeline Low Dose" ...
$ ACTARMCD: chr  "Pbo" "Pbo" "Xan_Hi" "Xan_Lo" ...
$ ACTARM  : chr  "Placebo" "Placebo" "Xanomeline High Dose" "Xanomeline Low Dose" ...
$ COUNTRY : chr  "USA" "USA" "USA" "USA" ...
$ DMDTC   : chr  "2013-12-26" "2012-07-22" "2013-07-11" "2014-03-10" ...
$ DMDY    : int  -7 -14 -8 -8 -7 -21 NA -9 -13 -7 ...
}
\usage{
dm
}
\description{
dm
}
\keyword{datasets}
