% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNTSBdf.R
\name{downloadNTSBdf}
\alias{downloadNTSBdf}
\title{Download NTSB Accidents and Incidents Database}
\source{
\url{https://app.ntsb.gov/aviationquery/}
}
\usage{
downloadNTSBdf()
}
\value{
A data.frame with one row per accident or incident and one column for each field the NTSB makes available.
}
\description{
Function to download a fresh copy of the NTSB accidents and incidents database from the web as a data.frame.
Note that a 2017-03-28 version of the NTSB database created with this function is available with the package, see NTSBData_20170328.
}
\examples{
df <- downloadNTSBdf() # Downloads and stores as .GlobalEnv object
}
