% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ds}
\alias{ds}
\title{ds}
\format{
a data frame with 596 rows and  13  columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ DSSEQ }}{ integer - Sequence Number }
\item{\code{ DSSPID }}{ text - Sponsor-Defined Identifier }
\item{\code{ DSTERM }}{ text - Reported Term for the Disposition Event }
\item{\code{ DSDECOD }}{ text - Standardized Disposition Term }
\item{\code{ DSCAT }}{ text - Category for Disposition Event }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ DSDTC }}{ datetime - Date/Time of Collection }
\item{\code{ DSSTDTC }}{ date - Start Date/Time of Disposition Event }
\item{\code{ DSSTDY }}{ integer - Study Day of Start of Disposition Event }
}
}
\usage{
ds
}
\description{
Disposition  data frame - One record per disposition status or protocol milestone per subject
}
\keyword{datasets}
