% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAS_to_TAS.R
\name{EAS_to_TAS}
\alias{EAS_to_TAS}
\title{Convert Equivalent Airspeed to True Airspeed}
\source{
US Navy, FAA (1965-01-02)
Aerodynamics for Naval Aviators (NAVAIR/NAVWEPS 00-80T-80)
\url{https://www.faa.gov/regulations_policies/handbooks_manuals/aviation/}

NACA, ICAO, NASA (Published 1954-05-01, Acquired 1996-09-01)
Manual of the ICAO standard atmosphere calculations by the NACA (NASA Technical Reports Server Document ID 19930083952)
\url{https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19930083952.pdf}
}
\usage{
EAS_to_TAS(EAS_kts, Altimeter_inHg = 29.92, Temp_C = 15)
}
\arguments{
\item{EAS_kts}{Equivalent airspeed in knots}

\item{Altimeter_inHg}{Local altimeter setting measured as inches of mercury, defaults to 29.92 (ICAO standard atmosphere)}

\item{Temp_C}{Local outside air temperature in degrees Celcius, defaults to 15 (ICAO standard temperature)}
}
\value{
True airspeed in knots
}
\description{
Function to convert aircraft's equivalent airspeed (EAS) to true airspeed (TAS).
TAS is equal to EAS adjusted for non-standard atmospheric pressure and temperature.
For modern commercial aircraft that opearte below the speed of sound, indicated airspeed (IAS) and EAS are practically equal to each other.
They may be used interchangeably in the context of measuring aircraft speeds within these assuptions.
}
\examples{
EAS_to_TAS(EAS_kts = 100, Altimeter_inHg = 29.80, Temp_C = 35)
EAS_to_TAS(EAS_kts = 100) # At standard temperature and pressure, EAS is almost exactly the same as TAS
}
\seealso{
\code{\link{TAS_to_GS}}
}
