% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ex}
\alias{ex}
\title{ex}
\format{
data.frame
'data.frame':	591 obs. of  17 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "EX" "EX" "EX" "EX" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1023" ...
$ EXSEQ   : int  1 2 3 1 2 1 2 3 1 1 ...
$ EXTRT   : chr  "PLACEBO" "PLACEBO" "PLACEBO" "PLACEBO" ...
$ EXDOSE  : int  0 0 0 0 0 54 81 54 54 54 ...
$ EXDOSU  : chr  "mg" "mg" "mg" "mg" ...
$ EXDOSFRM: chr  "PATCH" "PATCH" "PATCH" "PATCH" ...
$ EXDOSFRQ: chr  "QD" "QD" "QD" "QD" ...
$ EXROUTE : chr  "TRANSDERMAL" "TRANSDERMAL" "TRANSDERMAL" "TRANSDERMAL" ...
$ VISITNUM: int  3 4 12 3 4 3 4 12 3 3 ...
$ VISIT   : chr  "BASELINE" "WEEK 2" "WEEK 24" "BASELINE" ...
$ VISITDY : int  1 14 168 1 14 1 14 168 1 1 ...
$ EXSTDTC : chr  "2014-01-02" "2014-01-17" "2014-06-19" "2012-08-05" ...
$ EXENDTC : chr  "2014-01-16" "2014-06-18" "2014-07-02" "2012-08-27" ...
$ EXSTDY  : int  1 16 169 1 24 1 15 173 1 1 ...
$ EXENDY  : int  15 168 182 23 28 14 172 180 14 15 ...
}
\usage{
ex
}
\description{
ex
}
\keyword{datasets}
