% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{lb}
\alias{lb}
\title{lb}
\format{
data.frame
'data.frame':	59580 obs. of  23 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "LB" "LB" "LB" "LB" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ LBSEQ   : int  1 39 74 104 134 164 199 229 259 294 ...
$ LBTESTCD: chr  "ALB" "ALB" "ALB" "ALB" ...
$ LBTEST  : chr  "Albumin" "Albumin" "Albumin" "Albumin" ...
$ LBCAT   : chr  "CHEMISTRY" "CHEMISTRY" "CHEMISTRY" "CHEMISTRY" ...
$ LBORRES : chr  "3.8" "3.9" "3.8" "3.7" ...
$ LBORRESU: chr  "g/dL" "g/dL" "g/dL" "g/dL" ...
$ LBORNRLO: num  3.3 3.3 3.3 3.3 3.3 3.3 3.3 3.3 3.3 3.3 ...
$ LBORNRHI: num  4.9 4.9 4.9 4.9 4.9 4.9 4.9 4.9 4.9 4.9 ...
$ LBSTRESC: chr  "38" "39" "38" "37" ...
$ LBSTRESN: num  38 39 38 37 38 38 37 37 38 38 ...
$ LBSTRESU: chr  "g/L" "g/L" "g/L" "g/L" ...
$ LBSTNRLO: num  33 33 33 33 33 33 33 33 33 33 ...
$ LBSTNRHI: num  49 49 49 49 49 49 49 49 49 49 ...
$ LBNRIND : chr  "NORMAL" "NORMAL" "NORMAL" "NORMAL" ...
$ LBBLFL  : chr  "Y" NA NA NA ...
$ VISITNUM: num  1 4 5 7 8 9 10 11 12 13 ...
$ VISIT   : chr  "SCREENING 1" "WEEK 2" "WEEK 4" "WEEK 6" ...
$ VISITDY : int  -7 14 28 42 56 84 112 140 168 182 ...
$ LBDTC   : chr  "2013-12-26T14:45" "2014-01-16T13:17" "2014-01-30T08:50" "2014-02-12T12:56" ...
$ LBDY    : int  -7 15 29 42 63 84 126 140 168 182 ...
}
\usage{
lb
}
\description{
lb
}
\keyword{datasets}
