% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{mh}
\alias{mh}
\title{mh}
\format{
data.frame
'data.frame':	1818 obs. of  19 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "MH" "MH" "MH" "MH" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ MHSEQ   : int  9 1 8 2 10 3 4 6 11 7 ...
$ MHSPID  : chr  NA "E01" NA "E03" ...
$ MHTERM  : chr  "ALZHEIMER'S DISEASE" "VERBATIM_0135" "VERBATIM_0140" "VERBATIM_0301" ...
$ MHLLT   : chr  NA "PALPITATIONS" "SUBTOTAL HYSTERECTOMY" "HEADACHE" ...
$ MHDECOD : chr  NA "PALPITATIONS" "HYSTERECTOMY" "HEADACHE" ...
$ MHHLT   : chr  NA "HLT_0493" "HLT_0675" "HLT_0064" ...
$ MHHLGT  : chr  NA "HLGT_0593" "HLGT_0657" "HLGT_0344" ...
$ MHCAT   : chr  "PRIMARY DIAGNOSIS" "SIGNIFICANT PRE-EXISTING CONDITION" "HISTORICAL DIAGNOSIS" "SIGNIFICANT PRE-EXISTING CONDITION" ...
$ MHBODSYS: chr  NA "CARDIAC DISORDERS" "SURGICAL AND MEDICAL PROCEDURES" "NERVOUS SYSTEM DISORDERS" ...
$ MHSEV   : chr  NA "MILD" NA "MILD" ...
$ VISITNUM: int  1 1 1 1 1 1 1 1 1 1 ...
$ VISIT   : chr  "SCREENING 1" "SCREENING 1" "SCREENING 1" "SCREENING 1" ...
$ VISITDY : int  -7 -7 -7 -7 -7 -7 -7 -7 -7 -7 ...
$ MHDTC   : chr  "2013-12-26" "2013-12-26" "2013-12-26" "2013-12-26" ...
$ MHSTDTC : chr  "2010-04-30" NA "1986" NA ...
$ MHDY    : int  -7 -7 -7 -7 -7 -7 -7 -7 -7 -7 ...
}
\usage{
mh
}
\description{
mh
}
\keyword{datasets}
