% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NTSBData_20170328}
\alias{NTSBData_20170328}
\title{NTSB Accidents and Incidents from 2017-03-28}
\format{A data frame with 79557 rows and 31 variables:
\describe{
  \item{Event.Id}{Unique NTSB event ID}
  \item{Investigation.Type}{Factor indicating accident or incident}
  ...
}}
\source{
\url{https://app.ntsb.gov/aviationquery/}
}
\usage{
NTSBData_20170328
}
\description{
A copy of the NTSB accidents and incidents database from 2017-03-28.
Created using \code{\link{downloadNTSBdf}}
}
\keyword{datasets}
