% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_abline_intersect.R
\name{qq_abline_intersect}
\alias{qq_abline_intersect}
\title{y-Intercept of Theoretical Normal Distribution}
\usage{
qq_abline_intersect(x, slope = NULL)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer corresponding to a y-intercept
}
\description{
Function to calculate the y-intercept of the theoretical normal distribution for a vector of numbers.
This can be used to visually compare actual vs. theoretical values to make an interpretation on whether a sample of numbers is
likely driven from a normally distributed population. Calculations are largely inspired from \code{\link{stats::qqline}} function.
This can be used with \code{\link{ggplot2::geom_qq}} to create a Q-Q plot.

Note: Requires a slope to be provided, otherwise uses \code{\link{qq_abline_slope}} to calculate it (but will not return slope).
}
\examples{
qq_abline_intersect(1:10)
}
\seealso{
\code{\link{qq_abline_slope}}

\code{\link{ggplot2::geom_qq}}

\code{\link{stats::qqline}}
}
