% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_abline_slope.R
\name{qq_abline_slope}
\alias{qq_abline_slope}
\title{Slope of Theoretical Normal Distribution}
\usage{
qq_abline_slope(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer corresponding to a slope
}
\description{
Function to calculate the slope of the theoretical normal distribution for a vector of numbers.
This can be used to visually compare actual vs. theoretical values to make an interpretation on whether a sample of numbers is
likely driven from a normally distributed population. Calculations are largely inspired from \code{\link{stats::qqline}} function.
This can be used with \code{\link{ggplot2::geom_qq}} to create a Q-Q plot.
}
\examples{
qq_abline_slope(1:10)
}
\seealso{
\code{\link{qq_abline_intersect}}

\code{\link{ggplot2::geom_qq}}

\code{\link{stats::qqline}}
}
