% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{qs}
\alias{qs}
\title{qs}
\format{
data.frame
'data.frame':	121749 obs. of  20 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "QS" "QS" "QS" "QS" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ QSSEQ   : int  5001 5016 5031 5046 5002 5017 5032 5047 5003 5018 ...
$ QSTESTCD: chr  "ACITM01" "ACITM01" "ACITM01" "ACITM01" ...
$ QSTEST  : chr  "WORD RECALL TASK" "WORD RECALL TASK" "WORD RECALL TASK" "WORD RECALL TASK" ...
$ QSCAT   : chr  "ALZHEIMER'S DISEASE ASSESSMENT SCALE" "ALZHEIMER'S DISEASE ASSESSMENT SCALE" "ALZHEIMER'S DISEASE ASSESSMENT SCALE" "ALZHEIMER'S DISEASE ASSESSMENT SCALE" ...
$ QSSCAT  : chr  NA NA NA NA ...
$ QSORRES : chr  "3" "2" "4" "3" ...
$ QSORRESU: chr  NA NA NA NA ...
$ QSSTRESC: num  3 2 4 3 1 0 0 0 3 3 ...
$ QSSTRESN: num  3 2 4 3 1 0 0 0 3 3 ...
$ QSSTRESU: chr  NA NA NA NA ...
$ QSBLFL  : chr  "Y" NA NA NA ...
$ QSDRVFL : chr  NA NA NA NA ...
$ VISITNUM: num  3 8 10 12 3 8 10 12 3 8 ...
$ VISIT   : chr  "BASELINE" "WEEK 8" "WEEK 16" "WEEK 24" ...
$ VISITDY : int  1 56 112 168 1 56 112 168 1 56 ...
$ QSDTC   : chr  "2014-01-02" "2014-03-05" "2014-05-07" "2014-06-18" ...
$ QSDY    : int  1 63 126 168 1 63 126 168 1 63 ...
}
\usage{
qs
}
\description{
qs
}
\keyword{datasets}
