% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_events.R
\name{rare_events}
\alias{rare_events}
\title{Rare Events Search}
\usage{
rare_events(x, y)
}
\arguments{
\item{x}{eFOQA username in single quotes}

\item{y}{eFOQA password in single quotes}
}
\value{
Data frame consisting of flight record, flight date, tail number, fleet, takeoff airport code, landing airport code, and event type
}
\description{
Function to create data frame of rare events that need to be validated in EMS
}
