% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_events_metadata.R
\name{rare_events_metadata}
\alias{rare_events_metadata}
\title{Metadata for Rare Events Search}
\usage{
rare_events_metadata(x, y)
}
\arguments{
\item{x}{eFOQA username in single quotes}

\item{y}{eFOQA password in single quotes}
}
\value{
Db file (rare_events2.db) saved in the user's working directory
}
\description{
Function to create db file referenced in rare_events function
}
