% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{relrec}
\alias{relrec}
\title{relrec}
\format{
data.frame
'data.frame':	234 obs. of  7 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ RDOMAIN : chr  "AE" "AE" "AE" "AE" ...
$ USUBJID : chr  "01-701-1023" "01-701-1047" "01-701-1111" "01-701-1115" ...
$ IDVAR   : chr  "AESEQ" "AESEQ" "AESEQ" "AESEQ" ...
$ IDVARVAL: int  2 4 7 7 6 8 9 1 7 9 ...
$ RELTYPE : logi  NA NA NA NA NA NA ...
$ RELID   : chr  "01-701-1023-E09" "01-701-1047-E09" "01-701-1111-E16" "01-701-1115-E13" ...
}
\usage{
relrec
}
\description{
relrec
}
\keyword{datasets}
