% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{sc}
\alias{sc}
\title{sc}
\format{
a data frame with 254 rows and  14  columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ SCSEQ }}{ integer - Sequence Number }
\item{\code{ SCTESTCD }}{ text - Subject Characteristic Short Name }
\item{\code{ SCTEST }}{ text - Subject Characteristic }
\item{\code{ SCCAT }}{ text - Category for Subject Characteristic }
\item{\code{ SCORRES }}{ text - Result or Finding in Original Units }
\item{\code{ SCORRESU }}{ text - Original Units }
\item{\code{ SCSTRESC }}{ text - Character Result/Finding in Std Format }
\item{\code{ SCSTRESN }}{ integer - Numeric Result/Finding in Standard Units }
\item{\code{ SCSTRESU }}{ text - Standard Units }
\item{\code{ SCDTC }}{ date - Date/Time of Collection }
\item{\code{ SCDY }}{ integer - Study Day of Examination }
}
}
\usage{
sc
}
\description{
Subject Characteristics  data frame - One record per characteristic per subject
}
\keyword{datasets}
