% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_ae}
\alias{sdtm_ae}
\title{sdtm_ae}
\format{
a data frame with 1191 rows and 35 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ AESEQ }}{ integer - Sequence Number }
\item{\code{ AESPID }}{ text - Sponsor-Defined Identifier }
\item{\code{ AETERM }}{ text - Reported Term for the Adverse Event }
\item{\code{ AELLT }}{ text - Lowest Level Term }
\item{\code{ AELLTCD }}{ integer - Lowest Level Term Code }
\item{\code{ AEDECOD }}{ text - Dictionary-Derived Term }
\item{\code{ AEPTCD }}{ integer - Preferred Term Code }
\item{\code{ AEHLT }}{ text - High Level Term }
\item{\code{ AEHLTCD }}{ integer - High Level Term Code }
\item{\code{ AEHLGT }}{ text - High Level Group Term }
\item{\code{ AEHLGTCD }}{ integer - High Level Group Term Code }
\item{\code{ AEBODSYS }}{ text - Body System or Organ Class }
\item{\code{ AEBDSYCD }}{ integer - Body System or Organ Class Code }
\item{\code{ AESOC }}{ text - Primary System Organ Class }
\item{\code{ AESOCCD }}{ integer - Primary System Organ Class Code }
\item{\code{ AESEV }}{ text - Severity/Intensity }
\item{\code{ AESER }}{ text - Serious Event }
\item{\code{ AEACN }}{ text - Action Taken with Study Treatment }
\item{\code{ AEREL }}{ text - Causality }
\item{\code{ AEOUT }}{ text - Outcome of Adverse Event }
\item{\code{ AESCAN }}{ text - Involves Cancer }
\item{\code{ AESCONG }}{ text - Congenital Anomaly or Birth Defect }
\item{\code{ AESDISAB }}{ text - Persist or Signif Disability/Incapacity }
\item{\code{ AESDTH }}{ text - Results in Death }
\item{\code{ AESHOSP }}{ text - Requires or Prolongs Hospitalization }
\item{\code{ AESLIFE }}{ text - Is Life Threatening }
\item{\code{ AESOD }}{ text - Occurred with Overdose }
\item{\code{ AEDTC }}{ date - Date/Time of Collection }
\item{\code{ AESTDTC }}{ date - Start Date/Time of Adverse Event }
\item{\code{ AEENDTC }}{ date - End Date/Time of Adverse Event }
\item{\code{ AESTDY }}{ integer - Study Day of Start of Adverse Event }
\item{\code{ AEENDY }}{ integer - Study Day of End of Adverse Event }
}
}
\usage{
sdtm_ae
}
\description{
Adverse Events data frame - One record per adverse event per subject
}
\keyword{datasets}
