% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_cm}
\alias{sdtm_cm}
\title{sdtm_cm}
\format{
a data frame with 7510 rows and 21 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ CMSEQ }}{ integer - Sequence Number }
\item{\code{ CMSPID }}{ text - Sponsor-Defined Identifier }
\item{\code{ CMTRT }}{ text - Reported Name of Drug, Med, or Therapy }
\item{\code{ CMDECOD }}{ text - Standardized Medication Name }
\item{\code{ CMINDC }}{ text - Indication }
\item{\code{ CMCLAS }}{ text - Medication Class }
\item{\code{ CMDOSE }}{ integer - Dose per Administration }
\item{\code{ CMDOSU }}{ text - Dose Units }
\item{\code{ CMDOSFRQ }}{ text - Dosing Frequency per Interval }
\item{\code{ CMROUTE }}{ text - Route of Administration }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ CMDTC }}{ date - Date/Time of Collection }
\item{\code{ CMSTDTC }}{ date - Start Date/Time of Medication }
\item{\code{ CMENDTC }}{ date - End Date/Time of Medication }
\item{\code{ CMSTDY }}{ integer - Study Day of Start of Medication }
\item{\code{ CMENDY }}{ integer - Study Day of End of Medication }
}
}
\usage{
sdtm_cm
}
\description{
Concomitant Medications data frame - One record per recorded medication occurrence or constant-dosing interval per subject
}
\keyword{datasets}
