% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_dm}
\alias{sdtm_dm}
\title{sdtm_dm}
\format{
a data frame with 306 rows and 25 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ SUBJID }}{ text - Subject Identifier for the Study }
\item{\code{ RFSTDTC }}{ date - Subject Reference Start Date/Time }
\item{\code{ RFENDTC }}{ date - Subject Reference End Date/Time }
\item{\code{ RFXSTDTC }}{ datetime - Date/Time of First Study Treatment }
\item{\code{ RFXENDTC }}{ datetime - Date/Time of Last Study Treatment }
\item{\code{ RFICDTC }}{ datetime - Date/Time of Informed Consent }
\item{\code{ RFPENDTC }}{ datetime - Date/Time of End of Participation }
\item{\code{ DTHDTC }}{ datetime - Date/Time of Death }
\item{\code{ DTHFL }}{ text - Subject Death Flag }
\item{\code{ SITEID }}{ text - Study Site Identifier }
\item{\code{ AGE }}{ integer - Age }
\item{\code{ AGEU }}{ text - Age Units }
\item{\code{ SEX }}{ text - Sex }
\item{\code{ RACE }}{ text - Race }
\item{\code{ ETHNIC }}{ text - Ethnicity }
\item{\code{ ARMCD }}{ text - Planned Arm Code }
\item{\code{ ARM }}{ text - Description of Planned Arm }
\item{\code{ ACTARMCD }}{ text - Actual Arm Code }
\item{\code{ ACTARM }}{ text - Description of Actual Arm }
\item{\code{ COUNTRY }}{ text - Country }
\item{\code{ DMDTC }}{ date - Date/Time of Collection }
\item{\code{ DMDY }}{ integer - Study Day of Collection }
}
}
\usage{
sdtm_dm
}
\description{
Demographics data frame - One record per subject
}
\keyword{datasets}
