% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_ex}
\alias{sdtm_ex}
\title{sdtm_ex}
\format{
a data frame with 591 rows and 17 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ EXSEQ }}{ integer - Sequence Number }
\item{\code{ EXTRT }}{ text - Name of Actual Treatment }
\item{\code{ EXDOSE }}{ integer - Dose per Administration }
\item{\code{ EXDOSU }}{ text - Dose Units }
\item{\code{ EXDOSFRM }}{ text - Dose Form }
\item{\code{ EXDOSFRQ }}{ text - Dosing Frequency per Interval }
\item{\code{ EXROUTE }}{ text - Route of Administration }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ EXSTDTC }}{ date - Start Date/Time of Treatment }
\item{\code{ EXENDTC }}{ date - End Date/Time of Treatment }
\item{\code{ EXSTDY }}{ integer - Study Day of Start of Treatment }
\item{\code{ EXENDY }}{ integer - Study Day of End of Treatment }
}
}
\usage{
sdtm_ex
}
\description{
Exposure data frame - One record per constant dosing interval per subject
}
\keyword{datasets}
