% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_lb}
\alias{sdtm_lb}
\title{sdtm_lb}
\format{
a data frame with 59580 rows and 23 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ LBSEQ }}{ integer - Sequence Number }
\item{\code{ LBTESTCD }}{ text - Lab Test or Examination Short Name }
\item{\code{ LBTEST }}{ text - Lab Test or Examination Name }
\item{\code{ LBCAT }}{ text - Category for Lab Test }
\item{\code{ LBORRES }}{ text - Result or Finding in Original Units }
\item{\code{ LBORRESU }}{ text - Original Units }
\item{\code{ LBORNRLO }}{ text - Reference Range Lower Limit in Orig Unit }
\item{\code{ LBORNRHI }}{ text - Reference Range Upper Limit in Orig Unit }
\item{\code{ LBSTRESC }}{ text - Character Result/Finding in Std Format }
\item{\code{ LBSTRESN }}{ integer - Numeric Result/Finding in Standard Units }
\item{\code{ LBSTRESU }}{ text - Standard Units }
\item{\code{ LBSTNRLO }}{ integer - Reference Range Lower Limit-Std Units }
\item{\code{ LBSTNRHI }}{ integer - Reference Range Upper Limit-Std Units }
\item{\code{ LBNRIND }}{ text - Reference Range Indicator }
\item{\code{ LBBLFL }}{ text - Baseline Flag }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ LBDTC }}{ datetime - Date/Time of Specimen Collection }
\item{\code{ LBDY }}{ integer - Study Day of Specimen Collection }
}
}
\usage{
sdtm_lb
}
\description{
Laboratory Tests Results data frame - One record per analyte per planned time point number per time point reference per visit per subject
}
\keyword{datasets}
