% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_mh}
\alias{sdtm_mh}
\title{sdtm_mh}
\format{
a data frame with 1818 rows and 19 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ MHSEQ }}{ integer - Sequence Number }
\item{\code{ MHSPID }}{ text - Sponsor-Defined Identifier }
\item{\code{ MHTERM }}{ text - Reported Term for the Medical History }
\item{\code{ MHLLT }}{ text - Lowest Level Term }
\item{\code{ MHDECOD }}{ text - Dictionary-Derived Term }
\item{\code{ MHHLT }}{ text - High Level Term }
\item{\code{ MHHLGT }}{ text - High Level Group Term }
\item{\code{ MHCAT }}{ text - Category for Medical History }
\item{\code{ MHBODSYS }}{ text - Body System or Organ Class }
\item{\code{ MHSEV }}{ text - Severity/Intensity }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ MHDTC }}{ date - Date/Time of History Collection }
\item{\code{ MHSTDTC }}{ date - Start Date/Time of Medical History Event }
\item{\code{ MHDY }}{ integer - Study Day of History Collection }
}
}
\usage{
sdtm_mh
}
\description{
Medical History data frame - One record per medical history event per subject
}
\keyword{datasets}
