% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_qs}
\alias{sdtm_qs}
\title{sdtm_qs}
\format{
a data frame with 121749 rows and 20 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ QSSEQ }}{ integer - Sequence Number }
\item{\code{ QSTESTCD }}{ text - Question Short Name }
\item{\code{ QSTEST }}{ text - Question Name }
\item{\code{ QSCAT }}{ text - Category of Question }
\item{\code{ QSSCAT }}{ text - Subcategory for Question }
\item{\code{ QSORRES }}{ text - Finding in Original Units }
\item{\code{ QSORRESU }}{ text - Original Units }
\item{\code{ QSSTRESC }}{ text - Character Result/Finding in Std Format }
\item{\code{ QSSTRESN }}{ integer - Numeric Finding in Standard Units }
\item{\code{ QSSTRESU }}{ text - Standard Units }
\item{\code{ QSBLFL }}{ text - Baseline Flag }
\item{\code{ QSDRVFL }}{ text - Derived Flag }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ QSDTC }}{ date - Date/Time of Finding }
\item{\code{ QSDY }}{ integer - Study Day of Finding }
}
}
\usage{
sdtm_qs
}
\description{
Questionnaires data frame - One record per questionnaire per question per time point per visit per subject
}
\keyword{datasets}
