% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_relrec}
\alias{sdtm_relrec}
\title{sdtm_relrec}
\format{
a data frame with 234 rows and 7 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ RDOMAIN }}{ text - Related Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ IDVAR }}{ text - Identifying Variable }
\item{\code{ IDVARVAL }}{ text - Identifying Variable Value }
\item{\code{ RELTYPE }}{ text - Relationship Type }
\item{\code{ RELID }}{ text - Relationship Identifier }
}
}
\usage{
sdtm_relrec
}
\description{
Related Records data frame - One record per related record, group of records or datasets
}
\keyword{datasets}
