% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_se}
\alias{sdtm_se}
\title{sdtm_se}
\format{
a data frame with 752 rows and 9 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ SESEQ }}{ integer - Sequence Number }
\item{\code{ ETCD }}{ text - Element Code }
\item{\code{ ELEMENT }}{ text - Description of Element }
\item{\code{ SESTDTC }}{ date - Start Date/Time of Element }
\item{\code{ SEENDTC }}{ date - End Date/Time of Element }
\item{\code{ SEUPDES }}{ text - Description of Unplanned Element }
}
}
\usage{
sdtm_se
}
\description{
Subject Elements data frame - One record per actual Element per subject
}
\keyword{datasets}
