% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_suppae}
\alias{sdtm_suppae}
\title{sdtm_suppae}
\format{
a data frame with 1191 rows and 10 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ RDOMAIN }}{ text - Related Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ IDVAR }}{ text - Identifying Variable }
\item{\code{ IDVARVAL }}{ text - Identifying Variable Value }
\item{\code{ QNAM }}{ text - Qualifier Variable Name }
\item{\code{ QLABEL }}{ text - Qualifier Variable Label }
\item{\code{ QVAL }}{ text - Data Value }
\item{\code{ QORIG }}{ text - Origin }
\item{\code{ QEVAL }}{ text - Evaluator }
}
}
\usage{
sdtm_suppae
}
\description{
Supplemental Qualifiers for AE data frame - One record per IDVAR, IDVARVAL, and QNAM value per subject
}
\keyword{datasets}
