% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_sv}
\alias{sdtm_sv}
\title{sdtm_sv}
\format{
a data frame with 3559 rows and 8 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ SVSTDTC }}{ date - Start Date/Time of Visit }
\item{\code{ SVENDTC }}{ date - End Date/Time of Visit }
}
}
\usage{
sdtm_sv
}
\description{
Subject Visits data frame - One record per actual visit per subject
}
\keyword{datasets}
