% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_ta}
\alias{sdtm_ta}
\title{sdtm_ta}
\format{
a data frame with 8 rows and 10 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ ARMCD }}{ text - Planned Arm Code }
\item{\code{ ARM }}{ text - Description of Planned Arm }
\item{\code{ TAETORD }}{ integer - Order of Element within Arm }
\item{\code{ ETCD }}{ text - Element Code }
\item{\code{ ELEMENT }}{ text - Description of Element }
\item{\code{ TABRANCH }}{ text - Branch }
\item{\code{ TATRANS }}{ text - Transition Rule }
\item{\code{ EPOCH }}{ text - Epoch }
}
}
\usage{
sdtm_ta
}
\description{
Trial Arms data frame - One record per planned Element per Arm
}
\keyword{datasets}
