% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_te}
\alias{sdtm_te}
\title{sdtm_te}
\format{
a data frame with 7 rows and 7 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ ETCD }}{ text - Element Code }
\item{\code{ ELEMENT }}{ text - Description of Element }
\item{\code{ TESTRL }}{ text - Rule for Start of Element }
\item{\code{ TEENRL }}{ text - Rule for End of Element }
\item{\code{ TEDUR }}{ text - Planned Duration of Element }
}
}
\usage{
sdtm_te
}
\description{
Trial Elements data frame - One record per planned Element
}
\keyword{datasets}
