% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_ti}
\alias{sdtm_ti}
\title{sdtm_ti}
\format{
a data frame with 31 rows and 6 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ IETESTCD }}{ text - Incl/Excl Criterion Short Name }
\item{\code{ IETEST }}{ text - Inclusion/Exclusion Criterion }
\item{\code{ IECAT }}{ text - Inclusion/Exclusion Category }
\item{\code{ TIRL }}{ text - Inclusion/Exclusion Criterion Rule }
}
}
\usage{
sdtm_ti
}
\description{
Trial Inclusion/ Exclusion Criteria data frame - One record per I/E criterion
}
\keyword{datasets}
