% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_tv}
\alias{sdtm_tv}
\title{sdtm_tv}
\format{
a data frame with 21 rows and 9 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ ARMCD }}{ text - Planned Arm Code }
\item{\code{ ARM }}{ text - Description of Planned Arm }
\item{\code{ TVSTRL }}{ text - Visit Start Rule }
\item{\code{ TVENRL }}{ text - Visit End Rule }
}
}
\usage{
sdtm_tv
}
\description{
Trial Visits data frame - One record per planned Visit per Arm
}
\keyword{datasets}
