% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_vs}
\alias{sdtm_vs}
\title{sdtm_vs}
\format{
a data frame with 29643 rows and 24 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ USUBJID }}{ text - Unique Subject Identifier }
\item{\code{ VSSEQ }}{ integer - Sequence Number }
\item{\code{ VSTESTCD }}{ text - Vital Signs Test Short Name }
\item{\code{ VSTEST }}{ text - Vital Signs Test Name }
\item{\code{ VSPOS }}{ text - Vital Signs Position of Subject }
\item{\code{ VSORRES }}{ text - Result or Finding in Original Units }
\item{\code{ VSORRESU }}{ text - Original Units }
\item{\code{ VSSTRESC }}{ text - Character Result/Finding in Std Format }
\item{\code{ VSSTRESN }}{ integer - Numeric Result/Finding in Standard Units }
\item{\code{ VSSTRESU }}{ text - Standard Units }
\item{\code{ VSSTAT }}{ text - Completion Status }
\item{\code{ VSLOC }}{ text - Location of Vital Signs Measurement }
\item{\code{ VSBLFL }}{ text - Baseline Flag }
\item{\code{ VISITNUM }}{ float - Visit Number }
\item{\code{ VISIT }}{ text - Visit Name }
\item{\code{ VISITDY }}{ integer - Planned Study Day of Visit }
\item{\code{ VSDTC }}{ date - Date/Time of Measurements }
\item{\code{ VSDY }}{ integer - Study Day of Vital Signs }
\item{\code{ VSTPT }}{ text - Planned Time Point Name }
\item{\code{ VSTPTNUM }}{ integer - Planned Time Point Number }
\item{\code{ VSELTM }}{ text - Planned Elapsed Time from Time Point Ref }
\item{\code{ VSTPTREF }}{ text - Time Point Reference }
}
}
\usage{
sdtm_vs
}
\description{
Vital Signs data frame - One record per vital sign measurement per time point per visit per subject
}
\keyword{datasets}
