% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{se}
\alias{se}
\title{se}
\format{
data.frame
'data.frame':	752 obs. of  9 variables:
$ STUDYID: chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN : chr  "SE" "SE" "SE" "SE" ...
$ USUBJID: chr  "01-701-1015" "01-701-1015" "01-701-1023" "01-701-1023" ...
$ SESEQ  : int  1 4 1 4 6 1 3 4 5 1 ...
$ ETCD   : chr  "SCRN" "PBO" "SCRN" "PBO" ...
$ ELEMENT: chr  "Screen" "Placebo" "Screen" "Placebo" ...
$ SESTDTC: chr  "2013-12-26" "2014-01-02" "2012-07-22" "2012-08-05" ...
$ SEENDTC: chr  "2014-01-02" "2014-07-02" "2012-08-05" "2013-02-18" ...
$ SEUPDES: chr  NA NA NA NA ...
}
\usage{
se
}
\description{
se
}
\keyword{datasets}
