% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcChart.R
\name{spcChart}
\alias{spcChart}
\title{Statistical Process Control np- or p-Chart Using ggplot2}
\usage{
spcChart(filename, title = NA, type = c("np", "p"), group = c("monthly",
  "quarterly"), enddate = floor_date(today(), "month") - 1,
  startdate = enddate - months(24) + 1)
}
\arguments{
\item{filename}{Name of CSV file on working directory to be consumed; expects specific formatting:
First column contains dates in the format "Mmm, YYYY"
Second column contains a count of events/issues/deficiencies
Third column contains the number of relevant operations (i.e. flights)
There should be no skipped months within dataset
Ordered from oldest date first (top) to newest date last (bottom)}

\item{title}{Optional plot title}

\item{type}{Chart type; accepts np or p, default np}

\item{group}{Date grouping for data to be plotted; accepts monthly or quarterly, default monthly}

\item{enddate}{Ending date ; accepts lubridate functions, default last day of previous month from current date}

\item{startdate}{Starting date for data to be plotted; accepts lubridate functions, default first day 24 months before current date}
}
\value{
SPC np-chart
}
\description{
Function to create statistical process control (SPC) np-chart or p-chart from aviation safety datasets.
Uses qcc:qcc() to generate SPC chart values and ggplot2 to generate the plot.
}
\examples{
spcChart("data.csv")
ggplot2::ggsave(filename = "np-Chart.png", plot = spcChart(filename = "file.csv"), device = "png", path = getwd(), height = 7, width = 14, units = "in")
}
