% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{suppae}
\alias{suppae}
\title{suppae}
\format{
data.frame
'data.frame':	1191 obs. of  10 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ RDOMAIN : chr  "AE" "AE" "AE" "AE" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1023" ...
$ IDVAR   : chr  "AESEQ" "AESEQ" "AESEQ" "AESEQ" ...
$ IDVARVAL: int  1 2 3 1 2 3 4 1 2 1 ...
$ QNAM    : chr  "AETRTEM" "AETRTEM" "AETRTEM" "AETRTEM" ...
$ QLABEL  : chr  "TREATMENT EMERGENT FLAG" "TREATMENT EMERGENT FLAG" "TREATMENT EMERGENT FLAG" "TREATMENT EMERGENT FLAG" ...
$ QVAL    : chr  "Y" "Y" "Y" "Y" ...
$ QORIG   : chr  "DERIVED" "DERIVED" "DERIVED" "DERIVED" ...
$ QEVAL   : chr  "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" ...
}
\usage{
suppae
}
\description{
suppae
}
\keyword{datasets}
