% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{suppdm}
\alias{suppdm}
\title{suppdm}
\format{
data.frame
'data.frame':	1197 obs. of  10 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ RDOMAIN : chr  "DM" "DM" "DM" "DM" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ IDVAR   : logi  NA NA NA NA NA NA ...
$ IDVARVAL: logi  NA NA NA NA NA NA ...
$ QNAM    : chr  "COMPLT16" "COMPLT24" "COMPLT8" "EFFICACY" ...
$ QLABEL  : chr  "Completers of Week 16 Population Flag" "Completers of Week 24 Population Flag" "Completers of Week 8 Population Flag" "Efficacy Population Flag" ...
$ QVAL    : chr  "Y" "Y" "Y" "Y" ...
$ QORIG   : chr  "DERIVED" "DERIVED" "DERIVED" "DERIVED" ...
$ QEVAL   : chr  "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" ...
}
\usage{
suppdm
}
\description{
suppdm
}
\keyword{datasets}
