% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{suppds}
\alias{suppds}
\title{suppds}
\format{
data.frame
'data.frame':	3 obs. of  10 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01"
$ RDOMAIN : chr  "DS" "DS" "DS"
$ USUBJID : chr  "01-703-1175" "01-705-1382" "01-708-1372"
$ IDVAR   : chr  "DSSEQ" "DSSEQ" "DSSEQ"
$ IDVARVAL: int  1 1 1
$ QNAM    : chr  "ENTCRIT" "ENTCRIT" "ENTCRIT"
$ QLABEL  : chr  "PROTOCOL ENTRY CRITERIA NOT MET" "PROTOCOL ENTRY CRITERIA NOT MET" "PROTOCOL ENTRY CRITERIA NOT MET"
$ QVAL    : int  16 25 16
$ QORIG   : chr  "CRF" "CRF" "CRF"
$ QEVAL   : logi  NA NA NA
}
\usage{
suppds
}
\description{
suppds
}
\keyword{datasets}
