% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{supplb}
\alias{supplb}
\title{supplb}
\format{
data.frame
'data.frame':	64403 obs. of  10 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ RDOMAIN : chr  "LB" "LB" "LB" "LB" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ IDVAR   : chr  "LBSEQ" "LBSEQ" "LBSEQ" "LBSEQ" ...
$ IDVARVAL: int  1 2 3 5 6 7 8 9 10 11 ...
$ QNAM    : chr  "LBTMSHI" "LBTMSHI" "LBTMSHI" "LBTMSHI" ...
$ QLABEL  : chr  "LAB RESULT/UPPER LIMIT OF NORMAL" "LAB RESULT/UPPER LIMIT OF NORMAL" "LAB RESULT/UPPER LIMIT OF NORMAL" "LAB RESULT/UPPER LIMIT OF NORMAL" ...
$ QVAL    : chr  "0.8" "0.3" "0.8" "1.2" ...
$ QORIG   : chr  "DERIVED" "DERIVED" "DERIVED" "DERIVED" ...
$ QEVAL   : chr  "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" "CLINICAL STUDY SPONSOR" ...
}
\usage{
supplb
}
\description{
supplb
}
\keyword{datasets}
