% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ta}
\alias{ta}
\title{ta}
\format{
data.frame
'data.frame':	8 obs. of  10 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "TA" "TA" "TA" "TA" ...
$ ARMCD   : chr  "Pbo" "Pbo" "Xan_Hi" "Xan_Hi" ...
$ ARM     : chr  "Placebo" "Placebo" "Xanomeline High Dose" "Xanomeline High Dose" ...
$ TAETORD : int  1 2 1 2 3 4 1 2
$ ETCD    : chr  "SCRN" "PBO" "SCRN" "HIS" ...
$ ELEMENT : chr  "Screen" "Placebo" "Screen" "High_Start" ...
$ TABRANCH: chr  "Randomized to Placebo" NA "Randomized to High Dose" NA ...
$ TATRANS : logi  NA NA NA NA NA NA ...
$ EPOCH   : chr  "Screening" "Treatment" "Screening" "Treatment" ...
}
\usage{
ta
}
\description{
ta
}
\keyword{datasets}
