% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAS_to_GS.R
\name{TAS_to_GS}
\alias{TAS_to_GS}
\title{Convert True Airspeed to Ground Speed}
\usage{
TAS_to_GS(TAS_kts, Direct_Wind_Component_kts = 0)
}
\arguments{
\item{TAS_kts}{True airspeed in knots}

\item{Direct_Wind_Component_kts}{Direct wind component in knots (headwind = positive; tailwind = negative), defaults to 0}
}
\value{
Groundspeed in knots
}
\description{
Function to convert aircraft's true airspeed (TAS) to groundspeed (GS).
GS is equal to TAS adjusted for direct wind.
}
\examples{
TAS_to_GS(TAS_kts = 100) # TAS == GS when direct wind is 0kts
TAS_to_GS(TAS_kts = 100, 10) # Headwind reduces GS
TAS_to_GS(TAS_kts = 100, -10) # Tailwind increases GS
}
\seealso{
\code{\link{EAS_to_TAS}}

\code{\link{windVectors_headwind}}
}
