% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{te}
\alias{te}
\title{te}
\format{
data.frame
'data.frame':	7 obs. of  7 variables:
$ STUDYID: chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN : chr  "TE" "TE" "TE" "TE" ...
$ ETCD   : chr  "FOLO" "HIE" "HIM" "HIS" ...
$ ELEMENT: chr  "Follow_up" "High_End" "High_Middle" "High_Start" ...
$ TESTRL : chr  "End of last scheduled visit on study (including early termination)" "Administration of first dose (from patches supplied at Visit 12)" "Administration of first dose (from patches supplied at Visit 4)" "Administration of first dose" ...
$ TEENRL : chr  "Completion of all specified followup activities (which vary on a patient-by-patient basis)" NA NA NA ...
$ TEDUR  : chr  NA "P2W" "P22W" "P2W" ...
}
\usage{
te
}
\description{
te
}
\keyword{datasets}
