% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ti}
\alias{ti}
\title{ti}
\format{
data.frame
'data.frame':	31 obs. of  6 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "TI" "TI" "TI" "TI" ...
$ IETESTCD: chr  "INCL01" "INCL02" "INCL03" "INCL04" ...
$ IETEST  : chr  "Males and postmenopausal females at least 50 years of age." "Diagnosis of probable AD as defined by NINCDS and the ADRDA guidelines (Protocol Attachment LZZT.7)." "MMSE score of 10 to 23." "Modified Hachinski Ischemic Scale score of <= 4. (Protocol Attachment LZZT.8)." ...
$ IECAT   : chr  "INCLUSION" "INCLUSION" "INCLUSION" "INCLUSION" ...
$ TIRL    : logi  NA NA NA NA NA NA ...
}
\usage{
ti
}
\description{
ti
}
\keyword{datasets}
