% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ts}
\alias{ts}
\title{ts}
\format{
data.frame
'data.frame':	33 obs. of  6 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "TS" "TS" "TS" "TS" ...
$ TSSEQ   : int  1 1 1 1 2 1 1 1 1 1 ...
$ TSPARMCD: chr  "ADDON" "AGEMAX" "AGEMIN" "AGESPAN" ...
$ TSPARM  : chr  "Added on to Existing Treatments" "Planned Maximum Age of Subjects" "Planned Minimum Age of Subjects" "Age Group" ...
$ TSVAL   : chr  "Y" "No maximum" "50 years" "ADULT (18-65)" ...
}
\usage{
ts
}
\description{
ts
}
\keyword{datasets}
