% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{tv}
\alias{tv}
\title{tv}
\format{
data.frame
'data.frame':	21 obs. of  9 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "TV" "TV" "TV" "TV" ...
$ VISITNUM: num  1 2 3 3.5 4 5 6 7 8 8.1 ...
$ VISIT   : chr  "SCREENING 1" "SCREENING 2" "BASELINE" "AMBUL ECG PLACEMENT" ...
$ VISITDY : int  -7 -1 1 13 14 28 30 42 56 70 ...
$ ARMCD   : logi  NA NA NA NA NA NA ...
$ ARM     : logi  NA NA NA NA NA NA ...
$ TVSTRL  : chr  "Start of Screen Epoch" "Start of Visitnum 3 - 24H" "(Just Before) Start of Treatment Epoch" "Start of Visitnum  4 - 24H" ...
$ TVENRL  : chr  "completion of screening activities" "subject leaves clinic after connection of ambulatory ECG machine" "After application of first dose" NA ...
}
\usage{
tv
}
\description{
tv
}
\keyword{datasets}
