% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USA_DaylightSavings_20170404}
\alias{USA_DaylightSavings_20170404}
\title{2010's USA Daylight Savings Dates from 2017-04-04}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{Change_Name}{Name of change; either ST -> DST ("spring forward") or DST -> ST ("fall back")}
  \item{Change_Value}{Integer representing changed time in hours (+1 or -1)}
  \item{Date_Time}{Local daylight savings time change date and time}
}}
\source{
\url{https://www.timeanddate.com/time/zone/usa/new-york}
}
\usage{
USA_DaylightSavings_20170404
}
\description{
A list of 2010-2019 USA daylight savings change dates from 2017-04-04.
}
\keyword{datasets}
