% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{vs}
\alias{vs}
\title{vs}
\format{
data.frame
'data.frame':	29643 obs. of  24 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "VS" "VS" "VS" "VS" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1015" "01-701-1015" ...
$ VSSEQ   : int  1 2 3 4 5 6 7 8 9 10 ...
$ VSTESTCD: chr  "DIABP" "DIABP" "DIABP" "DIABP" ...
$ VSTEST  : chr  "Diastolic Blood Pressure" "Diastolic Blood Pressure" "Diastolic Blood Pressure" "Diastolic Blood Pressure" ...
$ VSPOS   : chr  "SUPINE" "STANDING" "STANDING" "SUPINE" ...
$ VSORRES : num  64 83 57 68 59 71 56 51 61 67 ...
$ VSORRESU: chr  "mmHg" "mmHg" "mmHg" "mmHg" ...
$ VSSTRESC: num  64 83 57 68 59 71 56 51 61 67 ...
$ VSSTRESN: num  64 83 57 68 59 71 56 51 61 67 ...
$ VSSTRESU: chr  "mmHg" "mmHg" "mmHg" "mmHg" ...
$ VSSTAT  : chr  NA NA NA NA ...
$ VSLOC   : chr  NA NA NA NA ...
$ VSBLFL  : chr  NA NA NA NA ...
$ VISITNUM: num  1 1 1 2 2 2 3 3 3 3.5 ...
$ VISIT   : chr  "SCREENING 1" "SCREENING 1" "SCREENING 1" "SCREENING 2" ...
$ VISITDY : int  -7 -7 -7 -1 -1 -1 1 1 1 13 ...
$ VSDTC   : chr  "2013-12-26" "2013-12-26" "2013-12-26" "2013-12-31" ...
$ VSDY    : int  -7 -7 -7 -2 -2 -2 1 1 1 13 ...
$ VSTPT   : chr  "AFTER LYING DOWN FOR 5 MINUTES" "AFTER STANDING FOR 1 MINUTE" "AFTER STANDING FOR 3 MINUTES" "AFTER LYING DOWN FOR 5 MINUTES" ...
$ VSTPTNUM: int  815 816 817 815 816 817 815 816 817 815 ...
$ VSELTM  : chr  "PT5M" "PT1M" "PT3M" "PT5M" ...
$ VSTPTREF: chr  "PATIENT SUPINE" "PATIENT STANDING" "PATIENT STANDING" "PATIENT SUPINE" ...
}
\usage{
vs
}
\description{
vs
}
\keyword{datasets}
