% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windVectors_crosswind.R
\name{windVectors_crosswind}
\alias{windVectors_crosswind}
\title{Calculate Crosswind Component}
\usage{
windVectors_crosswind(True_Heading_deg, True_Wind_Direction_From_deg,
  Wind_Speed_kts)
}
\arguments{
\item{True_Heading_deg}{True aircraft heading in degrees of bearing (North == 0)}

\item{Wind_Speed_kts}{Wind speed in knots}

\item{True_Wind_Direction_From_Deg}{True wind direction (from) in degrees of bearing (North == 0)}
}
\value{
Crosswind component in knots (from right = positive; from left = negative)
}
\description{
Function to calculate crosswind component from total wind vector.
}
\examples{
windVectors_crosswind(0, 0, 10) # Direct headwind has no crosswind component
windVectors_crosswind(0, 90, 10) # Crosswind coming from the right has a positive value
windVectors_crosswind(0, 350, 10) # Crosswind coming from the left has a negative value
windVectors_crosswind(0, 630, 10) # Degree values >=360 can be used (360 == 0)
}
\seealso{
\code{\link{windVectors_crosswind}}
}
