% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windVectors_headwind.R
\name{windVectors_headwind}
\alias{windVectors_headwind}
\title{Calculate Direct Wind Component}
\usage{
windVectors_headwind(True_Heading_deg, True_Wind_Direction_From_deg,
  Wind_Speed_kts)
}
\arguments{
\item{True_Heading_deg}{True aircraft heading in degrees of bearing (North == 0)}

\item{Wind_Speed_kts}{Wind speed in knots}

\item{True_Wind_Direction_From_Deg}{True wind direction (from) in degrees of bearing (North == 0)}
}
\value{
Direct wind component in knots (headwind = positive; tailwind = negative)
}
\description{
Function to calculate direct wind component from total wind vector.
}
\examples{
windVectors_headwind(0, 270, 10) # Direct crosswind has no direct wind component
windVectors_headwind(0, 0, 10) # Headwind has a positive value
windVectors_headwind(0, 150, 10) # Tailwind has a negative value
windVectors_headwind(0, 540, 10) # Degree values >=360 can be used (360 == 0)
}
\seealso{
\code{\link{windVectors_crosswind}}
}
