% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{ds}
\alias{ds}
\title{ds}
\format{
data.frame
'data.frame':	596 obs. of  13 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "DS" "DS" "DS" "DS" ...
$ USUBJID : chr  "01-701-1015" "01-701-1015" "01-701-1023" "01-701-1023" ...
$ DSSEQ   : int  1 2 1 2 3 1 2 1 2 3 ...
$ DSSPID  : int  NA NA 24 NA NA NA NA NA NA NA ...
$ DSTERM  : chr  "PROTOCOL COMPLETED" "FINAL LAB VISIT" "ADVERSE EVENT" "FINAL LAB VISIT" ...
$ DSDECOD : chr  "COMPLETED" "FINAL LAB VISIT" "ADVERSE EVENT" "FINAL LAB VISIT" ...
$ DSCAT   : chr  "DISPOSITION EVENT" "OTHER EVENT" "DISPOSITION EVENT" "OTHER EVENT" ...
$ VISITNUM: num  13 13 5 5 201 13 13 5 5 201 ...
$ VISIT   : chr  "WEEK 26" "WEEK 26" "WEEK 4" "WEEK 4" ...
$ DSDTC   : chr  "2014-07-02" "2014-07-02T11:45" "2012-09-02" "2012-09-02T10:15" ...
$ DSSTDTC : chr  "2014-07-02" "2014-07-02" "2012-09-02" "2012-09-02" ...
$ DSSTDY  : int  182 182 29 29 198 180 180 28 28 182 ...
}
\usage{
ds
}
\description{
ds
}
\keyword{datasets}
