% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_events.R
\name{rare_events}
\alias{rare_events}
\title{Rare Event Search}
\usage{
rare_events(username, password, metadata.db = "event_validation.db",
  metadata.exists = FALSE, metadata.keep = FALSE, path = getwd())
}
\arguments{
\item{username}{eFOQA username}

\item{password}{eFOQA password}

\item{metadata.db}{file name of metadata}

\item{metadata.exists}{logical. Does the metadata file exist in the working directory?}

\item{metadata.keep}{logical. Should the metadata file be kept after running the function? Note: Due to permissioning, the metadata file may be kept even if metadata.keep = FALSE}

\item{path}{specify metadata file location for saving and retrieval}
}
\value{
events that require validation in EMS that can be saved as a data frame
}
\description{
A function used to determine rare events that require validation in EMS
}
\examples{
rare_events('username', 'password', \\cr
metadata.db = 'event_validation.db', \\cr
metadata.exists = FALSE, metadata.keep = FALSE, \\cr
path = getwd()))
}
