% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{sc}
\alias{sc}
\title{sc}
\format{
data.frame
'data.frame':	254 obs. of  14 variables:
$ STUDYID : chr  "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" "CDISCPILOT01" ...
$ DOMAIN  : chr  "SC" "SC" "SC" "SC" ...
$ USUBJID : chr  "01-701-1015" "01-701-1023" "01-701-1028" "01-701-1033" ...
$ SCSEQ   : int  1 1 1 1 1 1 1 1 1 1 ...
$ SCTESTCD: chr  "EDLEVEL" "EDLEVEL" "EDLEVEL" "EDLEVEL" ...
$ SCTEST  : chr  "EDUCATION LEVEL" "EDUCATION LEVEL" "EDUCATION LEVEL" "EDUCATION LEVEL" ...
$ SCCAT   : chr  "EDUCATION" "EDUCATION" "EDUCATION" "EDUCATION" ...
$ SCORRES : int  16 14 16 12 9 8 18 22 12 14 ...
$ SCORRESU: chr  "YEARS" "YEARS" "YEARS" "YEARS" ...
$ SCSTRESC: int  16 14 16 12 9 8 18 22 12 14 ...
$ SCSTRESN: int  16 14 16 12 9 8 18 22 12 14 ...
$ SCSTRESU: chr  "YEARS" "YEARS" "YEARS" "YEARS" ...
$ SCDTC   : chr  "2013-12-26" "2012-07-22" "2013-07-11" "2014-03-10" ...
$ SCDY    : int  -7 -14 -8 -8 -7 -21 -9 -13 -7 -13 ...
}
\usage{
sc
}
\description{
sc
}
\keyword{datasets}
