% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_ts}
\alias{sdtm_ts}
\title{sdtm_ts}
\format{
a data frame with 33 rows and 6 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/sdtm/README.md} for more details and full data specification. Columns:
\describe{
\item{\code{ STUDYID }}{ text - Study Identifier }
\item{\code{ DOMAIN }}{ text - Domain Abbreviation }
\item{\code{ TSSEQ }}{ integer - Sequence Number }
\item{\code{ TSPARMCD }}{ text - Trial Summary Parameter Short Name }
\item{\code{ TSPARM }}{ text - Trial Summary Parameter }
\item{\code{ TSVAL }}{ text - Parameter Value }
}
}
\usage{
sdtm_ts
}
\description{
Trial Summary data frame - One record per trial summary parameter value
}
\keyword{datasets}
